Plugins: move datasource specific meta out of the main meta type (#16803)
